requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.3
pytz==2023.3
APScheduler==3.6.3
requests==2.31.0
python-telegram-bot==13.7
